JCF: DUNE-DAQ/daq-cmake#156: update the __init__.py files to work with changes to daq_add_python_bindings
