--- python/appmodel/__init__.py
@@ -1,13 +1,0 @@
-from ._daq_appmodel_dal_py import *
-
-__all__= ['generate_modules']
-
-class UnknownGeneratorException(Exception):
-    pass
-
-def generate_modules(confdb, app, session):
-    
-    mods = smart_dap_application_generate_modules(confdb._obj, app.id, session.id)
-
-    return [confdb.get_dal(m.class_name, m.id) for m in mods]
-
